some further refactoring of code
